# Backend/api/openapi.json
{
	"openapi": "3.1.0",
	"info": { "title": "WebPractik", "version": "0.1.0" },
	"paths": {
		"/user/": {
			"post": {
				"tags": ["user"],
				"summary": "Create New User",
				"operationId": "create_new_user_user__post",
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": { "$ref": "#/components/schemas/UserCreate" }
						}
					}
				},
				"responses": {
					"200": {
						"description": "Successful Response",
						"content": {
							"application/json": {
								"schema": { "$ref": "#/components/schemas/ShowUser" }
							}
						}
					},
					"422": {
						"description": "Validation Error",
						"content": {
							"application/json": {
								"schema": { "$ref": "#/components/schemas/HTTPValidationError" }
							}
						}
					}
				}
			},
			"delete": {
				"tags": ["user"],
				"summary": "Delete User By Id",
				"operationId": "delete_user_by_id_user__delete",
				"security": [{ "OAuth2PasswordBearer": [] }],
				"parameters": [
					{
						"name": "id",
						"in": "query",
						"required": true,
						"schema": { "type": "string", "format": "uuid", "title": "Id" }
					}
				],
				"responses": {
					"200": {
						"description": "Successful Response",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{ "type": "string", "format": "uuid" },
										{ "type": "null" }
									],
									"title": "Response Delete User By Id User  Delete"
								}
							}
						}
					},
					"422": {
						"description": "Validation Error",
						"content": {
							"application/json": {
								"schema": { "$ref": "#/components/schemas/HTTPValidationError" }
							}
						}
					}
				}
			},
			"patch": {
				"tags": ["user"],
				"summary": "Update User By Id",
				"operationId": "update_user_by_id_user__patch",
				"parameters": [
					{
						"name": "id",
						"in": "query",
						"required": true,
						"schema": { "type": "string", "format": "uuid", "title": "Id" }
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": { "$ref": "#/components/schemas/UpdateUserRequest" }
						}
					}
				},
				"responses": {
					"200": {
						"description": "Successful Response",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{ "type": "string", "format": "uuid" },
										{ "type": "null" }
									],
									"title": "Response Update User By Id User  Patch"
								}
							}
						}
					},
					"422": {
						"description": "Validation Error",
						"content": {
							"application/json": {
								"schema": { "$ref": "#/components/schemas/HTTPValidationError" }
							}
						}
					}
				}
			}
		},
		"/login/": {
			"post": {
				"tags": ["login"],
				"summary": "Login For Acess Token",
				"operationId": "login_for_acess_token_login__post",
				"requestBody": {
					"content": {
						"application/x-www-form-urlencoded": {
							"schema": {
								"$ref": "#/components/schemas/Body_login_for_acess_token_login__post"
							}
						}
					},
					"required": true
				},
				"responses": {
					"200": {
						"description": "Successful Response",
						"content": {
							"application/json": {
								"schema": { "$ref": "#/components/schemas/Token" }
							}
						}
					},
					"422": {
						"description": "Validation Error",
						"content": {
							"application/json": {
								"schema": { "$ref": "#/components/schemas/HTTPValidationError" }
							}
						}
					}
				}
			}
		}
	},
	"components": {
		"schemas": {
			"Body_login_for_acess_token_login__post": {
				"properties": {
					"grant_type": {
						"anyOf": [
							{ "type": "string", "pattern": "^password$" },
							{ "type": "null" }
						],
						"title": "Grant Type"
					},
					"username": { "type": "string", "title": "Username" },
					"password": { "type": "string", "title": "Password" },
					"scope": { "type": "string", "title": "Scope", "default": "" },
					"client_id": {
						"anyOf": [{ "type": "string" }, { "type": "null" }],
						"title": "Client Id"
					},
					"client_secret": {
						"anyOf": [{ "type": "string" }, { "type": "null" }],
						"title": "Client Secret"
					}
				},
				"type": "object",
				"required": ["username", "password"],
				"title": "Body_login_for_acess_token_login__post"
			},
			"HTTPValidationError": {
				"properties": {
					"detail": {
						"items": { "$ref": "#/components/schemas/ValidationError" },
						"type": "array",
						"title": "Detail"
					}
				},
				"type": "object",
				"title": "HTTPValidationError"
			},
			"ShowUser": {
				"properties": {
					"user_id": { "type": "string", "format": "uuid", "title": "User Id" },
					"name": { "type": "string", "title": "Name" },
					"surname": { "type": "string", "title": "Surname" },
					"email": { "type": "string", "format": "email", "title": "Email" },
					"is_active": { "type": "boolean", "title": "Is Active" }
				},
				"type": "object",
				"required": ["user_id", "name", "surname", "email", "is_active"],
				"title": "ShowUser"
			},
			"Token": {
				"properties": {
					"access_token": { "type": "string", "title": "Access Token" },
					"token_type": { "type": "string", "title": "Token Type" }
				},
				"type": "object",
				"required": ["access_token", "token_type"],
				"title": "Token"
			},
			"UpdateUserRequest": {
				"properties": {
					"name": {
						"anyOf": [{ "type": "string" }, { "type": "null" }],
						"title": "Name"
					},
					"surname": {
						"anyOf": [{ "type": "string" }, { "type": "null" }],
						"title": "Surname"
					},
					"email": {
						"anyOf": [
							{ "type": "string", "format": "email" },
							{ "type": "null" }
						],
						"title": "Email"
					}
				},
				"type": "object",
				"title": "UpdateUserRequest"
			},
			"UserCreate": {
				"properties": {
					"name": { "type": "string", "title": "Name" },
					"surname": { "type": "string", "title": "Surname" },
					"email": { "type": "string", "format": "email", "title": "Email" },
					"password": { "type": "string", "title": "Password" },
					"role": { "type": "string", "title": "Role" }
				},
				"type": "object",
				"required": ["name", "surname", "email", "password", "role"],
				"title": "UserCreate"
			},
			"ValidationError": {
				"properties": {
					"loc": {
						"items": { "anyOf": [{ "type": "string" }, { "type": "integer" }] },
						"type": "array",
						"title": "Location"
					},
					"msg": { "type": "string", "title": "Message" },
					"type": { "type": "string", "title": "Error Type" }
				},
				"type": "object",
				"required": ["loc", "msg", "type"],
				"title": "ValidationError"
			}
		},
		"securitySchemes": {
			"OAuth2PasswordBearer": {
				"type": "oauth2",
				"flows": { "password": { "scopes": {}, "tokenUrl": "/login" } }
			}
		}
	}
}